Refactored tests to better support params
needed to be able to provide multiple sets of params to stash::install
to see how the behaviour is changed.

Tweaked the class dependency model so that stash at the top level,
manages the ordering of classes, rather than stash::config and stash::install
linking between each other.

This should also reduce the number of Before[] and Require[] metaparams.

see Puppet Anchor Pattern for details.